⚡ Bolt: Centralize shared environment logic into env_utils.py
