+++ env_utils.py
@@ -1,0 +1,31 @@
+import os
+from functools import lru_cache
+
+def _parse_env_line(line, env_dict):
+    line = line.strip()
+    if not line or line.startswith("#"):
+        return
+    if line.startswith("export "):
+        line = line[7:].strip()
+    # ⚡ Bolt Optimization: Use partition() over split() to avoid intermediate list allocation overhead
+    key, sep, val = line.partition("=")
+    if not sep:
+        return
+    env_dict[key] = val.strip("'\"")
+
+@lru_cache(maxsize=None)
+def _get_parsed_env_vars():
+    # ⚡ Bolt Optimization: Cache only the parsed variables from the file to prevent redundant IO reads, while keeping it safe from mutable dictionary cache poisoning
+    parsed_vars = {}
+    try:
+        with open("../email-security-pipeline/GH_TOKEN.env", "r") as f:
+            for line in f:
+                _parse_env_line(line, parsed_vars)
+    except FileNotFoundError:
+        pass
+    return parsed_vars
+
+def _load_gh_token_env():
+    env = os.environ.copy()
+    env.update(_get_parsed_env_vars())
+    return env
